# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST",
        "version": "1.0.0"
    },
    "paths": {
        "/api/customer/{id}": {
            "get": {
                "summary": "Récupérer un client par ID",
                "operationId": "getCustomerById",
                "parameters": [
                    {
                        "in": "path",
                        "required": true,
                        "name": "id",
                        "description": "L'ID du client que vous voulez récupérer",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erreur du token JWT"
                    },
                    "403": {
                        "description": "Vous n'êtes pas autorisé à accèder à cette ressource."
                    },
                    "404": {
                        "description": "Aucun client trouvé avec cet ID"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/customer": {
            "get": {
                "summary": "Liste des clients",
                "operationId": "getCustomers",
                "responses": {
                    "200": {
                        "description": "Tous les clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erreur du token JWT"
                    },
                    "404": {
                        "description": "Aucun client trouvé"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "summary": "Créer un nouveau client",
                "operationId": "createCustomer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstname": {
                                        "description": "Prénom du client",
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "lastname": {
                                        "description": "Nom du client",
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "description": "E-mail du client",
                                        "type": "string",
                                        "example": "john.doe@email.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe du client",
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de syntaxe"
                    },
                    "401": {
                        "description": "Erreur du token JWT"
                    },
                    "403": {
                        "description": "Accès interdit"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/product": {
            "get": {
                "summary": "Liste des produits",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Tous les produits",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun produit trouvé"
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "summary": "Récupérer un produit par ID",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "in": "path",
                        "required": true,
                        "name": "id",
                        "description": "L'ID du produit que vous voulez récupérer",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du produit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun produit trouvé avec cet ID"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "summary": "Liste des utilisateurs",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Tous les utilisateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erreur du token JWT"
                    },
                    "404": {
                        "description": "Aucun utilisateur trouvé"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "summary": "Créer un nouvel utilisateur",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstname": {
                                        "description": "Prénom de l'utilisateur",
                                        "type": "string",
                                        "example": "Saul"
                                    },
                                    "lastname": {
                                        "description": "Nom de l'utilisateur",
                                        "type": "string",
                                        "example": "Goodman"
                                    },
                                    "email": {
                                        "description": "E-mail de l'utilisateur",
                                        "type": "string",
                                        "example": "saul.goodman@email.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe de l'utilisateur",
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de syntaxe"
                    },
                    "401": {
                        "description": "Erreur du token JWT"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "get": {
                "summary": "Récupérer un utilisateur par ID",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "in": "path",
                        "required": true,
                        "name": "id",
                        "description": "L'ID de l'utilisateur que vous voulez récupérer",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erreur du token JWT"
                    },
                    "403": {
                        "description": "Vous n'êtes pas autorisé à accèder à cette ressource."
                    },
                    "404": {
                        "description": "Cet utilisateur n'existe pas"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "summary": "Supprimer un utilisateur par ID",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "in": "path",
                        "required": true,
                        "name": "id",
                        "description": "ID de l'utilisateur qui va être supprimé",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Utilisateur supprimé"
                    },
                    "401": {
                        "description": "Erreur du token JWT"
                    },
                    "403": {
                        "description": "Vous n'êtes pas autorisé à accèder à cette ressource."
                    },
                    "404": {
                        "description": "Cet utilisateur n'existe pas"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Customer": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "firstname": {
                        "type": "string",
                        "nullable": true
                    },
                    "lastname": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "password": {
                        "description": "The hashed password",
                        "type": "string",
                        "nullable": true
                    },
                    "company": {
                        "type": "string",
                        "nullable": true
                    }
                }
            },
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "brand": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "password": {
                        "description": "The hashed password",
                        "type": "string",
                        "nullable": true
                    },
                    "firstname": {
                        "type": "string",
                        "nullable": true
                    },
                    "lastname": {
                        "type": "string",
                        "nullable": true
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}